% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blankRemoval.R
\name{blankRemoval}
\alias{blankRemoval}
\title{Remove backgroud signals}
\usage{
blankRemoval(peaks_all, peaks_ctrl)
}
\arguments{
\item{peaks_all}{a formated feature table of all biological samples}

\item{peaks_ctrl}{a formated feature table of control sample only}
}
\value{
a feature table
}
\description{
Use a peak table of CTRL samples to filter out background signals which have formed valid feature groups during processing.
In the first instance, extact matches are sought between the CTRL peak table and QC peak table using feature ID's. Exact
matches are removed from the QC peak table if the fold change in raw intensity is less than 2.0 between CTRL and QC. In the second round
CTRL signals are located within the Rt and mz ranges of the QC peak table. Again, any matches with less than 2.0 fold change between CTRL and QC
are removed.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\seealso{
featureTable
}

