% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkQCRSD.R
\name{checkQCRSD}
\alias{checkQCRSD}
\title{Check QC-RSD}
\usage{
checkQCRSD(peak_table, info)
}
\arguments{
\item{peak_table}{a formatted peak table.}

\item{info}{a runinfo \code{data.frame}. Quality control samples shoule be denoted by \code{QC} in the
\code{class} column.}
}
\value{
a \code{QC} \code{data.frame}
}
\description{
Determine the Coefficent of variation (CV) / Relative Standard Deviation (RSD) for all available Quality Control (QC) samples
}
\author{
Tom Wilsoin \email{tpw2@aber.ac.uk}
}
\seealso{
formatTable

plot.CV
}

