% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRMSD.R
\name{checkRSMD}
\alias{checkRSMD}
\title{Find outlier runs}
\usage{
checkRSMD(xcms_set, info)
}
\arguments{
\item{xcms_set}{an \code{xcmsSet} which has had retention time correction applied}

\item{info}{a runinfo \code{data.frame}}
}
\value{
a \code{RMSD data.frame}
}
\description{
Use the root mean square deviation (RMSD) between raw and corrected retenion times to identify potential outlier samples
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

