% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileMS2.R
\name{compileMS2}
\alias{compileMS2}
\title{Build dd FT-MS2 List}
\usage{
compileMS2(filename, save = FALSE)
}
\arguments{
\item{filename}{a raw \code{.mzML} or \code{.mzXML} file containing FT-MS1 and dd-FT-MS2 spectra only.}
}
\value{
returns an object of \code{class} \code{ftms2}

The \code{ftms2} class is a list containing the following elements
 \itemize{
   \item{pMz} the precursor m/z value
   \item{pRt} the precursor retention time in minutes
   \item{pInt} the precursor intensity
   \item{spec} a matrix of MS2 spectra.
 }
}
\description{
Compile a list object of data dependant (dd) MS2 spectra acquired using FT-MS, which can be used further downstream
for matching to MS1 feature groups and peak annotation
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

