% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVplot.R
\name{CVplot}
\alias{CVplot}
\title{Plot CV}
\usage{
CVplot(peak_table, info)
}
\arguments{
\item{peak_table}{A formatted peak table.}

\item{info}{A runinfo \code{data.frame}. Quality control samples shoule be denoted by \code{QC} in the
\code{class} column.}
}
\value{
A histogram is saved to the current working directory
}
\description{
Plot the Coefficent of variation (%) for all available Quality Control (QC) samples
}
\author{
Tom Wilsoin \email{tpw2@aber.ac.uk}
}

