% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPDA.R
\name{extractPDA}
\alias{extractPDA}
\title{Extract PDA Data}
\usage{
extractPDA(filepath, savedir)
}
\arguments{
\item{filepath}{a valid filepath containing \code{.mzML} files}

\item{savedir}{location to save extracted PDA data to}
}
\description{
Extract the UV Diodearray data from a \code{.mzML} file and save into a much smaller \code{.RData} file
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

