% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTable.r
\name{formatTable}
\alias{formatTable}
\title{Format xcmsSet}
\usage{
formatTable(xcms_set, feat_ids = FALSE)
}
\arguments{
\item{xcms_set}{A \code{xcmsSet} containing peak group information}

\item{feat_ids}{A logical statement. If \code{TRUE} a \code{.csv} file is written to the working
directory containing a numeric identification number for each feature.}
}
\value{
A formatted peak matrix
}
\description{
Format xcmsSet into peak table ready for statistical analysis
}
\author{
Tom Wilson \email{tpw2@bert.ibers.aber.uk}
}

