% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoCheck.R
\name{infoCheck}
\alias{infoCheck}
\title{Check info file}
\usage{
infoCheck(filepath, info)
}
\arguments{
\item{filepath}{the filepath containing raw (\code{.mzML, .mzXML}) files}

\item{info}{a runinfo \code{data.frame}. See \url{https://github.com/wilsontom/LCPRO/wiki} for more details in runinfo files}
}
\value{
a runinfo \code{data.frame}
}
\description{
Check that the supplied \code{info} file contains all the nesscary information in the correct format
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

