% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAll.R
\name{loadAll}
\alias{loadAll}
\title{Load all xcmsSet's}
\usage{
loadAll(path)
}
\arguments{
\item{path}{directory containing xcmsSet's which were saved using \code{saveAll}}
}
\description{
Load all of the xcmsSet's in the specified directory.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\seealso{
\link{saveAll}
}

