% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakDetection.R
\name{peakDetection}
\alias{peakDetection}
\title{Peak Detection}
\usage{
peakDetection(filepath, info, rtcor = TRUE)
}
\arguments{
\item{filepath}{the filepath containing raw (\code{.mzML, .mzXML}) files}

\item{info}{a runinfo  \code{data.frame}}

\item{rtcor}{logical; if \code{TRUE} then OBIWARP retention time correction is applied}
}
\value{
a \code{xcmsSet}
}
\description{
This is a wrapper function for LC-HRMS peak detection using xcms::findPeaks.centwave. Values for parameters
such as \code{snthresh}, \code{ppm}, \code{peakwidth}, etc.. have been hardcoded to reflect typical and optimised
values for non-targeted profiling of human urine using Reverse Phase (RP) C18 chromatography and Orbitrap mass detection.
}
\details{
The wrapper performs the nesscary checks of meta data prior to peak detection and helps to set up a multi-threaded (parallel)
environment
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

