% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ftms2.R
\name{plot.ftms2}
\alias{plot.ftms2}
\title{Plot dd-FT-MS2 Spectra}
\usage{
\method{plot}{ftms2}(x, saveplot = FALSE)
}
\arguments{
\item{x}{an object of \code{ftms2} class}

\item{saveplot}{logical; if \code{TRUE} the plot is saved to the current working directory.}
}
\description{
Plot the m/z and relative intensity of individual MS2 spectra
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

