% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.QCCV.R
\name{plot.QCCV}
\alias{plot.QCCV}
\title{Plot QC-RSD}
\usage{
\method{plot}{QCCV}(x)
}
\arguments{
\item{x}{a \code{QCCV data.frame}}
}
\value{
a histogram
}
\description{
Plot a histogram of QC-RSD values
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

