% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RMSD.R
\name{plot.RMSD}
\alias{plot.RMSD}
\title{Plot Retention Time Deviation}
\usage{
\method{plot}{RMSD}(x)
}
\arguments{
\item{x}{a \code{RMSD data.frame}}
}
\value{
a scatter plot
}
\description{
Plot a scatter plot of root mean standard deviation between raw and corrected retention times for each sample.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

