% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ticCheck.R
\name{plot.ticCheck}
\alias{plot.ticCheck}
\title{Plot ticCheck}
\usage{
\method{plot}{ticCheck}(x, ...)
}
\arguments{
\item{x}{a \code{ticCheck} class}
}
\description{
Plot the TIC profile for a batch of samples
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

