% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRMSD.R
\name{plotRMSD}
\alias{plotRMSD}
\title{Plot the root mean square deviation (RMSD) in order to visualise the extent of applied retention time
correction in each sample.}
\usage{
plotRMSD(xcmsSet, injord, cls, saveplot = TRUE)
}
\arguments{
\item{xcmsSet}{an \code{xcmsSet} which has had retention time's corrected}

\item{injord}{a numeric vector of the injection order for all raw files. This vector must be
in file order, not numeric order}

\item{saveplot}{a logical statement. If \code{TRUE} the plot is saved as \code{.svg} to the current working directory}

\item{class}{a character vecotr of class information for all the raw files. This vecotr must be in file order}
}
\description{
Plot the root mean square deviation (RMSD) in order to visualise the extent of applied retention time
correction in each sample.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

