% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTIC.R
\name{plotTIC}
\alias{plotTIC}
\title{Check the TIC of all samples}
\usage{
plotTIC(filepath, injord, cls, saveplot = TRUE)
}
\arguments{
\item{filepath}{a filepath of \code{.mzML} or \code{.mzXML} files}

\item{injord}{a numeric vector of the injection order for all raw files. This vector must be
in file order, not numeric order}

\item{saveplot}{a logical statement. If \code{TRUE} the plot is saved as \code{.svg} to the current working directory}

\item{class}{a character vecotr of class information for all the raw files. This vecotr must be in file order}
}
\description{
Plot the sample total ion count (TIC) for all samples in an experiment/batch. #'
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

