% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAll.R
\name{saveAll}
\alias{saveAll}
\title{Save all xcmsSet's}
\usage{
saveAll(path)
}
\arguments{
\item{path}{directory where xcmsSet's will be saved to}
}
\description{
Save all of the xcmsSet's currently in the global environment (\code{.GlobalEnv}) into the specified directory.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\seealso{
\link{loadAll}
}

