% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDirTree.R
\name{setDirTree}
\alias{setDirTree}
\title{Create Project Directory Tree}
\usage{
setDirTree(filepath)
}
\arguments{
\item{filepath}{the top level directory for the LC-MS project}
}
\description{
Create a structure of directories which can be used alongside a the package workflow
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

