% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableQC.R
\name{variableQC}
\alias{variableQC}
\title{Summary QC report for variables}
\usage{
variableQC(xcms_set)
}
\arguments{
\item{xcms_set}{a xcms set with peak group information}
}
\value{
a \code{data.frame} giving a summary of variable "quality"
}
\description{
Create a small report table, which can be used an a heuristic evaluation of a processed feature matrix
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

