% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsRawSet.R
\name{xcmsRawSet}
\alias{xcmsRawSet}
\title{Parallel xcmsSet without Rmpi}
\usage{
xcmsRawSet(filepath, info, ppm, snthresh, peakwidth)
}
\arguments{
\item{filepath}{the filepath containing raw (\code{.mzML, .mzXML}) files}

\item{info}{a runinfo  \code{data.frame}}

\item{ppm}{a numeric value for maximum tolerated ppm deviation in consecutive scans in ppm}

\item{snthresh}{a numeric value for the signal to noise threshold}

\item{peakwidth}{numeric range (min,max) for chromatographic peak width in seconds}
}
\value{
a \code{xcmsSet}
}
\description{
This function is wrapper for using xcmsRaw across a FORK cluster to construct
an xcmsSet. By using a FORK cluster, which is easily intitialed using the \code{parallel}
especially useful for when running xcmsSet on large UNIX cluster's, where \code{Rmpi} continues
to be problematic.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

