% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsRawSetMF.R
\name{xcmsRawSetMF}
\alias{xcmsRawSetMF}
\title{Parallel xcmsSet (matchedFilter) without Rmpi}
\usage{
xcmsRawSetMF(filepath, info, fwhm, snthresh)
}
\arguments{
\item{filepath}{the filepath containing raw (\code{.mzML, .mzXML}) files}

\item{info}{a runinfo  \code{data.frame}}

\item{fwhm}{a numeric value for the full width at half maximum}

\item{snthresh}{a numeric value for the signal to noise threshold}

\item{peakwidth}{numeric range (min,max) for chromatographic peak width in seconds}
}
\value{
a \code{xcmsSet}
}
\description{
This function is wrapper for using xcmsRaw across a FORK cluster to construct
an xcmsSet. By using a FORK cluster, which is easily intitialed using the \code{parallel}
especially useful for when running xcmsSet on large UNIX cluster's, where \code{Rmpi} continues
to be problematic.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

