% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureTable.R
\name{featureTable}
\alias{featureTable}
\title{Format xcmsSet into Feature table}
\usage{
featureTable(xcms_set)
}
\arguments{
\item{xcms_set}{a \code{xcmsSet} containing peak group information}
}
\value{
A \code{data.frame} containing feature group details
}
\description{
Format xcmsSet into a feature summary table
}
\author{
Tom Wilson \email{tpw2@bert.ibers.aber.uk}
}

